socials/requirements.txt
moviepy>=1.0.3
pillow>=8.0.0
requests>=2.25.0 